1. just relocation of classes 2. clean up
