+++ app/src/main/java/io/ak1/pixsample/samples/FragmentSample.kt
@@ -65,6 +65,7 @@ class FragmentSample : AppCompatActivity() {
         else
             PixBus.onBackPressedEvent()
     }
+
 }
 
 class ResultsFragment(private val clickCallback: View.OnClickListener) : Fragment() {
