+++ app/src/main/res/layout/activity_main.xml
@@ -87,4 +87,9 @@
         android:text="@string/pix_activity_implementation_sample"
         android:textColor="@color/primary_color_pix" />
 
+    <Space
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_weight="1" />
+
 </LinearLayout>
