+++ pix/src/main/java/io/ak1/pix/preview/PreviewFragment.kt
@@ -1,4 +1,4 @@
-package io.ak1.pix.ui
+package io.ak1.pix.preview
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,13 +6,16 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.activity.OnBackPressedCallback
 import androidx.databinding.DataBindingUtil
+import androidx.fragment.app.Fragment
 import androidx.lifecycle.MutableLiveData
+import io.ak1.pix.PixActivity
 import io.ak1.pix.R
 import io.ak1.pix.databinding.FragmentPreviewBinding
 import io.ak1.pix.helpers.PixBus
 import io.ak1.pix.helpers.PixEventCallback
 import io.ak1.pix.helpers.deleteImage
-import io.ak1.pix.ui.image_video_pager.ImageVideoPagerAdapter
+import io.ak1.pix.helpers.setupScreen
+import io.ak1.pix.preview.pager.ImageVideoPagerAdapter
 import io.ak1.pix.utility.ARG_PARAM_PIX
 import io.ak1.pix.utility.IMG_PICKER
 
@@ -20,14 +23,23 @@ import io.ak1.pix.utility.IMG_PICKER
  * Created by Pritam Dasgupta on 27th December, 2021
  * */
 
-class PreviewFragment: PixBaseFragment() {
+class PreviewFragment : Fragment() {
     private var fragmentPreviewBinding: FragmentPreviewBinding? = null
-    //model to fetch the image(s) from arguments
-    private var uriList: PixEventCallback.Results? = null
-    private var isCameraClicked = false
 
-    //observer to check the image resource callback status
+    // model to fetch the image(s) from arguments
+    private var uriList: PixEventCallback.Results? = null
+    private var isPhotoFromCamera = false
+
+    // observer to check the image resource callback status
     private var loaderLD: MutableLiveData<Boolean> = MutableLiveData(false)
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        requireActivity().let {
+            it.setupScreen()
+            it.actionBar?.hide()
+        }
+    }
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -35,14 +47,14 @@ class PreviewFragment: PixBaseFragment() {
         savedInstanceState: Bundle?
     ): View? {
         uriList = arguments?.getParcelable(ARG_PARAM_PIX) ?: PixEventCallback.Results()
-        isCameraClicked = arguments?.getInt(IMG_PICKER) == 1 || arguments?.getInt(IMG_PICKER) == 3
+        isPhotoFromCamera = arguments?.getInt(IMG_PICKER) == 1 || arguments?.getInt(IMG_PICKER) == 3
         fragmentPreviewBinding =
             DataBindingUtil.inflate(inflater, R.layout.fragment_preview, container, false)
         requireActivity()
             .onBackPressedDispatcher
             .addCallback(viewLifecycleOwner, object : OnBackPressedCallback(true) {
                 override fun handleOnBackPressed() {
-                    if (isCameraClicked) {
+                    if (isPhotoFromCamera) {
                         deleteImagesVideos()
                     }
                     (activity as? PixActivity)?.navController?.navigateUp()
@@ -55,7 +67,7 @@ class PreviewFragment: PixBaseFragment() {
         super.onViewCreated(view, savedInstanceState)
         fragmentPreviewBinding?.cancelBtn?.setOnClickListener {
             //if images are not required/accepted, deleting images from internal as well as external storage.
-            if (isCameraClicked) {
+            if (isPhotoFromCamera) {
                 deleteImagesVideos()
             }
             (activity as? PixActivity)?.navController?.navigateUp()
@@ -74,7 +86,8 @@ class PreviewFragment: PixBaseFragment() {
         }
         loaderLD.observe(viewLifecycleOwner, {
             it?.apply {
-                fragmentPreviewBinding?.cancelBtn?.visibility = if (this) View.VISIBLE else View.GONE
+                fragmentPreviewBinding?.cancelBtn?.visibility =
+                    if (this) View.VISIBLE else View.GONE
                 fragmentPreviewBinding?.doneBtn?.visibility = if (this) View.VISIBLE else View.GONE
             }
         })
