+++ pix/src/main/res/layout/fragment_preview.xml
@@ -36,7 +36,7 @@
             android:layout_marginRight="12dp"
             tools:ignore="VectorDrawableCompat" />
 
-        <io.ak1.pix.ui.image_video_pager.ImageVideoPager
+        <io.ak1.pix.preview.pager.ImageVideoPager
             android:id="@+id/imageVideoPager"
             android:layout_width="0dp"
             android:layout_height="0dp"
